split out the agent tool code into multiple files
